Separate tcp responsibility in access point
