+++ Kortleser/CardReaderLibrary/CardReader.cs
@@ -8,53 +8,6 @@ using System.Threading.Tasks;
 namespace CardReaderLibrary;
 public class CardReader
 {
-    private TcpClient client;
-    private NetworkStream stream;
-    public event EventHandler<string>? LogMessage;
-
-    public CardReader(string serverAddress, int serverPort)
-    {
-        client = new TcpClient(serverAddress, serverPort);
-        stream = client.GetStream();
-    }
-
-    public async Task SendRequestAsync(string request)
-    {
-        byte[] requestBytes = Encoding.ASCII.GetBytes(request);
-        await stream.WriteAsync(requestBytes, 0, requestBytes.Length);
-        OnLogMessage("Request sent.");
-    }
-
-    public async Task<string> ReceiveResponseAsync()
-    {
-        byte[] responseBuffer = new byte[1024];
-        int bytesRead = await stream.ReadAsync(responseBuffer, 0, responseBuffer.Length);
-        string response = Encoding.ASCII.GetString(responseBuffer, 0, bytesRead);
-        return response;
-    }
-
-    public void CloseConnection()
-    {
-        stream.Close();
-        client.Close();
-        OnLogMessage("Connection closed.");
-    }
-
-
-
-
-    protected virtual void OnLogMessage(string message)
-    {
-        LogMessage?.Invoke(this, message);
-    }
-
-
-
-
-
-
-
-
     public void ForceMessage()
     {
         SendCommand("$R1");
